+++ @vates/diff/.npmignore
@@ -1,0 +1,1 @@
+../../scripts/npmignore
