+++ @vates/stream-reader/index.js
