+++ CHANGELOG.unreleased.md
@@ -34,7 +34,9 @@
 
 <!--packages-start-->
 
+- @vates/diff minor
 - @vates/read-chunk patch
+- @vates/stream-reader minor
 - @xen-orchestra/backups minor
 - vhd-lib minor
 - xo-web patch
